backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.2.1
alembic==1.13.1
python-multipart==0.0.9
email-validator==2.1.1